Seal a bunch of IR types to get better warnings.
When a match is on a sealed type, the compiler has enough information to
warn when that match is not exhaustive (i.e. you forgot a case).
