+++ src/main/scala/firrtl/passes/Passes.scala
@@ -38,9 +38,9 @@ import scala.collection.mutable.ArrayBuffer
 
 import firrtl._
 import firrtl.ir._
+import firrtl.ir.PrimOps._
 import firrtl.Utils._
 import firrtl.Mappers._
-import firrtl.PrimOps._
 import firrtl.WrappedExpression._
 
 trait Pass extends LazyLogging {
